# parts: don't consider parts overlapping if a.end == b.start